pyproject.toml
[project]
name = "ctrl-apply"
version = "0.1.0"
description = "Job application automation tool — Chrome extension + Python backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "claude-agent-sdk>=0.1.27",
    "playwright>=1.49.0",
    "playwright-stealth>=1.0.6",
    "python-jobspy>=1.1.75",
    "pymupdf>=1.25.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "duckdb>=1.2.0",
    "polars>=1.20.0",
    "rapidfuzz>=3.11.0",
    "pyyaml>=6.0.2",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
